Move display/completer.py to _internal/
