+++ certbot/_internal/display/completer.py
