+++ certbot/tests/display/completer_test.py
@@ -1,4 +1,4 @@
-"""Test certbot.display.completer."""
+"""Test certbot._internal.display.completer."""
 try:
     import readline  # pylint: disable=import-error
 except ImportError:
@@ -18,7 +18,7 @@ import certbot.tests.util as test_util  # pylint: disable=ungrouped-imports
 
 
 class CompleterTest(test_util.TempDirTestCase):
-    """Test certbot.display.completer.Completer."""
+    """Test certbot._internal.display.completer.Completer."""
 
     def setUp(self):
         super(CompleterTest, self).setUp()
@@ -40,7 +40,7 @@ class CompleterTest(test_util.TempDirTestCase):
                     pass
 
     def test_complete(self):
-        from certbot.display import completer
+        from certbot._internal.display import completer
         my_completer = completer.Completer()
         num_paths = len(self.paths)
 
@@ -64,7 +64,7 @@ class CompleterTest(test_util.TempDirTestCase):
         sys.modules['readline'] = original_readline
 
     def test_context_manager_with_unmocked_readline(self):
-        from certbot.display import completer
+        from certbot._internal.display import completer
         reload_module(completer)
 
         original_completer = readline.get_completer()
@@ -76,18 +76,18 @@ class CompleterTest(test_util.TempDirTestCase):
         self.assertEqual(readline.get_completer(), original_completer)
         self.assertEqual(readline.get_completer_delims(), original_delims)
 
-    @mock.patch('certbot.display.completer.readline', autospec=True)
+    @mock.patch('certbot._internal.display.completer.readline', autospec=True)
     def test_context_manager_libedit(self, mock_readline):
         mock_readline.__doc__ = 'libedit'
         self._test_context_manager_with_mock_readline(mock_readline)
 
-    @mock.patch('certbot.display.completer.readline', autospec=True)
+    @mock.patch('certbot._internal.display.completer.readline', autospec=True)
     def test_context_manager_readline(self, mock_readline):
         mock_readline.__doc__ = 'GNU readline'
         self._test_context_manager_with_mock_readline(mock_readline)
 
     def _test_context_manager_with_mock_readline(self, mock_readline):
-        from certbot.display import completer
+        from certbot._internal.display import completer
 
         mock_readline.parse_and_bind.side_effect = enable_tab_completion
